added infix
getting ready for messing around with cobra. needed the infix style to
go with the already supported post and pre fix styles that are already
supported. next step will be to add cmd line support using cobra. also,
add the feature that empty values and vectors can be passed into
MakeVectors and NewValues.